Fix compilation on MacOS (#121)

+++ src/cunumeric/binary/binary_op_util.h
@@ -911,6 +911,6 @@ struct RHS2OfBinaryOp<BinaryOpCode::LDEXP, CODE> {
 template <BinaryOpCode OP_CODE, legate::Type::Code CODE>
 using rhs2_of_binary_op = typename RHS2OfBinaryOp<OP_CODE, CODE>::type;
 
-std::vector<size_t> broadcast_shapes(std::vector<NDArray> arrays);
+std::vector<uint64_t> broadcast_shapes(std::vector<NDArray> arrays);
 
 }  // namespace cunumeric
